% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbulate.R
\name{carbulate}
\alias{carbulate}
\title{Calculate Carbonate Speciation using DIC, pH, and Temperature}
\usage{
carbulate(df, DIC_col_mg.L, pH_col, temp_col_C, pressure_col_kPa)
}
\arguments{
\item{df}{Your dataframe with parameters of interest}

\item{DIC_col_mg.L}{Name of column that contains your measured dissolved inorganic carbon concentration (mgC/L)}

\item{pH_col}{Name of column with your measured pH}

\item{temp_col_C}{Name of column with your measured water temperature (in celsius)}

\item{pressure_kPa}{Name of column with the field atmospheric pressure (in kPa)}
}
\description{
This function calculates the speciation of dissolved inorganic carbon given the measured pH, DIC concentration, and water temperature.
It will add on the concenrations of each carbonate species (CO2(aq), HCO3-, CO3--) to your dataframe 'df'.
}
\examples{
SAMMS_17 <- carbulate(SAMMS_17, 'DIC_mgC.L', 'pH', 'Temp_C')
}
\author{
Pieter J. K. Aukes
}
\keyword{DIC,}
\keyword{carbonate}
\keyword{speciation}
